add replicasets.yaml and services.yaml
