+++ tests/test-deployment/dump/replicasets.yaml
@@ -1,0 +1,75 @@
+apiVersion: v1
+items:
+- apiVersion: extensions/v1beta1
+  kind: ReplicaSet
+  metadata:
+    annotations:
+      deployment.kubernetes.io/desired-replicas: "3"
+      deployment.kubernetes.io/max-replicas: "4"
+      deployment.kubernetes.io/revision: "1"
+    creationTimestamp: "2019-09-25T12:02:51Z"
+    generation: 1
+    labels:
+      app: redis
+      pod-template-hash: 7f86d8599d
+      role: master
+      tier: backend
+    name: redis-master-7f86d8599d
+    namespace: default
+    ownerReferences:
+    - apiVersion: apps/v1
+      blockOwnerDeletion: true
+      controller: true
+      kind: Deployment
+      name: redis-master
+      uid: 65efde42-df8c-11e9-bb74-42010a8001fc
+    resourceVersion: "2234348"
+    selfLink: /apis/extensions/v1beta1/namespaces/default/replicasets/redis-master-7f86d8599d
+    uid: 65f15959-df8c-11e9-bb74-42010a8001fc
+  spec:
+    replicas: 3
+    selector:
+      matchLabels:
+        app: redis
+        pod-template-hash: 7f86d8599d
+        role: master
+        tier: backend
+    template:
+      metadata:
+        creationTimestamp: null
+        labels:
+          app: redis
+          pod-template-hash: 7f86d8599d
+          role: master
+          tier: backend
+      spec:
+        containers:
+        - image: k8s.gcr.io/redis:e2e
+          imagePullPolicy: IfNotPresent
+          name: master
+          ports:
+          - containerPort: 6379
+            protocol: TCP
+          resources:
+            limits:
+              memory: 500Mi
+            requests:
+              cpu: 100m
+              memory: 500Mi
+          terminationMessagePath: /dev/termination-log
+          terminationMessagePolicy: File
+        dnsPolicy: ClusterFirst
+        restartPolicy: Always
+        schedulerName: default-scheduler
+        securityContext: {}
+        terminationGracePeriodSeconds: 30
+  status:
+    availableReplicas: 3
+    fullyLabeledReplicas: 3
+    observedGeneration: 1
+    readyReplicas: 3
+    replicas: 3
+kind: List
+metadata:
+  resourceVersion: ""
+  selfLink: ""
